Restructure kubeless integration test folder (#671)
* Restructure kubeless integration folder
