+++ prow/jobs/kyma/tests/end-to-end/kubeless-integration/kubeless-integration.yaml
@@ -16,6 +16,25 @@ job_template: &job_template
         command:
           - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"
         args:
+          - "/home/prow/go/src/github.com/kyma-project/kyma/tests/end-to-end/kubeless-integration"
+        resources:
+          requests:
+            memory: 1.5Gi
+            cpu: 0.8
+
+job_template_rel06_rel07: &job_template_rel06_rel07
+  skip_report: false
+  decorate: true
+  path_alias: github.com/kyma-project/kyma
+  max_concurrency: 10
+  spec:
+    containers:
+      - image: eu.gcr.io/kyma-project/prow/test-infra/buildpack-golang:v20181119-afd3fbd
+        securityContext:
+          privileged: true
+        command:
+          - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"
+        args:
           - "/home/prow/go/src/github.com/kyma-project/kyma/tests/kubeless-integration"
         resources:
           requests:
@@ -35,30 +54,30 @@ presubmits: # runs on PRs
       <<: *job_template
       run_if_changed: "^tests/kubeless-integration/"
       extra_refs:
-      - <<: *test_infra_ref
-        base_ref: master
+        - <<: *test_infra_ref
+          base_ref: master
       labels:
         <<: *job_labels_template
         preset-build-pr: "true"
     - name: pre-rel06-kyma-tests-kubeless-integration
       branches:
-      - release-0.6
-      <<: *job_template
+        - release-0.6
+      <<: *job_template_rel06_rel07
       always_run: true
       extra_refs:
-      - <<: *test_infra_ref
-        base_ref: release-0.6
+        - <<: *test_infra_ref
+          base_ref: release-0.6
       labels:
         <<: *job_labels_template
         preset-build-release: "true"
     - name: pre-rel07-kyma-tests-kubeless-integration
       branches:
-      - release-0.7
-      <<: *job_template
+        - release-0.7
+      <<: *job_template_rel06_rel07
       always_run: true
       extra_refs:
-      - <<: *test_infra_ref
-        base_ref: release-0.7
+        - <<: *test_infra_ref
+          base_ref: release-0.7
       labels:
         <<: *job_labels_template
         preset-build-release: "true"
@@ -71,8 +90,8 @@ postsubmits:
       <<: *job_template
       run_if_changed: "^tests/kubeless-integration/"
       extra_refs:
-      - <<: *test_infra_ref
-        base_ref: master
+        - <<: *test_infra_ref
+          base_ref: master
       labels:
         <<: *job_labels_template
         preset-build-master: "true"
